Update Client to handle 400 responses from the controller
