+++ app.go
@@ -30,10 +30,10 @@ var appNamePattern = regexp.MustCompile(`^[a-z\d]+(-[a-z\d]+)*$`)
 func (r *AppRepo) Add(data interface{}) error {
 	app := data.(*ct.App)
 	if app.Name == "" {
-		return ValidationError{"name", "must not be blank"}
+		return ct.ValidationError{"name", "must not be blank"}
 	}
 	if len(app.Name) > 30 || !appNamePattern.MatchString(app.Name) {
-		return ValidationError{"name", "is invalid"}
+		return ct.ValidationError{"name", "is invalid"}
 	}
 	if app.ID == "" {
 		app.ID = utils.UUID()
