+++ types/types.go
@@ -2,6 +2,7 @@ package types
 
 import (
 	"encoding/json"
+	"fmt"
 	"time"
 )
 
@@ -106,4 +107,13 @@ type ResourceReq struct {
 	ProviderID string           `json:"-"`
 	Apps       []string         `json:"apps,omitempty"`
 	Config     *json.RawMessage `json:"config"`
+}
+
+type ValidationError struct {
+	Field   string `json:"field,omitempty"`
+	Message string `json:"message"`
+}
+
+func (v ValidationError) Error() string {
+	return fmt.Sprintf("validation error: %s %s", v.Field, v.Message)
 }
